renamed feature changes as described in the pull requests comments, except the sphinx docs
